// PatientPortal/src/app/patient-history/models/occupational-history.model.ts
export class OccupationalHistoryModel {
    id: string;
    occupationalType: string;
    start: any;
    end: any;
    disabilityClaimDetails: string;
    workersCompensationClaimDetails: string;
    employmentStatus: string;
    notes: string;
    patientId: string;
    createDate: any;

    constructor() {
        this.createDate = new Date();
    }
}